requirements.txt
torch
torchvision
transformers
pytorch-lightning
timm
supervision
pycocotools
pillow
scipy
matplotlib